// parcial/ClientApp/src/app/Apoyos/models/persona.ts
export class Persona {
        identificacion: string;
        nombres: string;
        apellidos: string;
        sexo: string;
        edad: number;
        departamento: string;
        ciudad: string;
        valorApoyo: number;
        modalidad: string;
        fecha: string;
    }
    